+++ docs/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Improve
 
+* Fix arrow key bug in hangul [#373](https://github.com/Riey/kime/issues/373)
+
 ## 2.0.0
 
 ### Breaking